Move abstract code into @annotator/selector
