+++ packages/dom/src/text-node-chunker.ts
@@ -18,58 +18,9 @@
  * under the License.
  */
 
+import type { Chunk, Chunker, ChunkRange } from '@annotator/selector';
 import { normalizeRange } from './normalize-range';
 import { ownerDocument } from './owner-document';
-
-// A Chunk represents a fragment (typically a string) of some document.
-// Subclasses can add further attributes to map the chunk to its position in the
-// data structure it came from (e.g. a DOM node).
-export interface Chunk<TData> {
-  readonly data: TData;
-  equals?(otherChunk: this): boolean;
-}
-
-export interface ChunkRange<TChunk extends Chunk<any>> {
-  startChunk: TChunk;
-  startIndex: number;
-  endChunk: TChunk;
-  endIndex: number;
-}
-
-export function chunkEquals(chunk1: Chunk<any>, chunk2: Chunk<any>): boolean {
-  return chunk1.equals ? chunk1.equals(chunk2) : chunk1 === chunk2;
-}
-
-export function chunkRangeEquals(
-  range1: ChunkRange<any>,
-  range2: ChunkRange<any>,
-): boolean {
-  return (
-    chunkEquals(range1.startChunk, range2.startChunk) &&
-    chunkEquals(range1.endChunk, range2.endChunk) &&
-    range1.startIndex === range2.startIndex &&
-    range1.endIndex === range2.endIndex
-  );
-}
-
-// A Chunker lets one walk through the chunks of a document.
-// It is inspired by, and similar to, the DOM’s NodeIterator. (but unlike
-// NodeIterator, it has no concept of being ‘before’ or ‘after’ a chunk)
-export interface Chunker<TChunk extends Chunk<any>> {
-  // The chunk currently being pointed at.
-  readonly currentChunk: TChunk;
-
-  // Move currentChunk to the chunk following it, and return that chunk.
-  // If there are no chunks following it, keep currentChunk unchanged and return null.
-  nextChunk(): TChunk | null;
-
-  // Move currentChunk to the chunk preceding it, and return that chunk.
-  // If there are no preceding chunks, keep currentChunk unchanged and return null.
-  previousChunk(): TChunk | null;
-
-  // Test if a given chunk is before the current chunk.
-  precedesCurrentChunk(chunk: TChunk): boolean;
-}
 
 export interface PartialTextNode extends Chunk<string> {
   readonly node: Text;
