+++ packages/selector/src/text/code-point-seeker.ts
