+++ packages/selector/src/text/seek.ts
